Added session to pages

+++ delete.php
@@ -1,14 +1,24 @@
 <?php
+session_start();
 
 include "top.php";
-include 'DBHandler.php';
 
-$dbh = new DBHandler('SQL', 'mysql', 3306, 'root', 'docker', 'MikesMovies');
-$dbh->connect();
-// $dbh->setSQLConnection($dbh->getDBConnection());
+if (@$_SESSION)
+{
+    include 'DBHandler.php';
 
-$dbh->delete($_GET["id"]);
+    $dbh = new DBHandler('SQL', 'mysql', 3306, 'root', 'docker', 'MikesMovies');
+    $dbh->connect();
+    // $dbh->setCRUDConnection($dbh->getDBConnection());
+    
+    $dbh->delete($_GET["id"]);
+    
+    echo '<a href="index.php">Back</a>';
+}
 
-echo '<a href="index.php">Back</a>';
+else 
+{
+    include "index.php";
+}
 
 ?>
